Crate af_opencl_interop

af-opencl-interop package is to used only when the application intends to mix arrayfire code with raw OpenCL code.

Functions from this crate return OpenCL C API opaque pointers typedefs

cl_device_id

cl_context

cl_command_queue

Enums

DeviceType
OpenCL Device Type

VendorPlatform
OpenCL Vendor Platform

Functions

add_device_context
Push user provided device, context and queue tuple to ArrayFire device mamanger

delete_device_context
Remove the user provided device, context pair from ArrayFire device mamanger

get_context
Get the handle to active ArrayFire OpenCL context

get_device_id
Get the handle to active ArrayFire OpenCL device

get_device_type

get_platform
Fetch Active ArrayFire device's vendor platform

get_queue
Get the handle to active ArrayFire OpenCL command queue

set_device_context
Set the device identified by device & context pair as the active device for ArrayFire

set_device_id
Set the cl_device_id as the active ArrayFire OpenCL device